moved markdown code to core folder
